[BOD-320] Refactor on Anchor/Button logic
